files moved and reorganized, now to hook it all up...
