+++ code/apps/archives/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class ArchivesConfig(AppConfig):
+    name = 'archives'
+
+    def ready(self):
+        import archives.signals
+
