+++ code/apps/simulations/static/js/general_assembler.js
