+++ code/apps/simulations/static/js/simulator.js
