+++ code/apps/simulations/templates/base_simulation.jinja2
