+++ code/apps/simulations/templates/floating_point_operations.jinja2
