+++ code/apps/simulations/templates/general_assembler.jinja2
