+++ code/apps/simulations/templates/looping_with_tix.jinja2
