+++ code/config/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for djweb project.
+Django settings for the dh@mit computational history project.
 
 Generated by 'django-admin startproject' using Django 2.1.
 
@@ -12,11 +12,12 @@ https://docs.djangoproject.com/en/2.1/ref/settings/
 
 import os
 from django_jinja.builtins import DEFAULT_EXTENSIONS
-
-from .jinja_utils import collect_jinja2_functions
+from utilities.jinja_utils import collect_jinja2_functions
 
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
+PROJECT_ROOT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+DJANGO_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 
@@ -48,8 +49,8 @@ INSTALLED_APPS = [
     'django_jinja.contrib._humanize',
 
     # our apps
-    'dj_comp_hist.apps.DjCompHistConfig',
-    'simulations.apps.SimulationsConfig',
+    'apps.archives',
+    'apps.simulations',
 ]
 
 MIDDLEWARE = [
@@ -62,7 +63,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'djweb.urls'
+ROOT_URLCONF = 'config.urls'
 
 TEMPLATES = [
     {
@@ -109,7 +110,7 @@ JINJA2_ENVIRONMENT_OPTIONS = {
 }
 JINJA2_FILTERS = collect_jinja2_functions()  # gets functions and filters
 
-WSGI_APPLICATION = 'djweb.wsgi.application'
+WSGI_APPLICATION = 'config.wsgi.application'
 
 
 # Database
