+++ code/utilities/metadata_parser.py
