--- djweb/dj_comp_hist/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'dj_comp_hist.apps.DjCompHistConfig'
