--- djweb/dj_comp_hist/apps.py
@@ -1,9 +1,0 @@
-from django.apps import AppConfig
-
-
-class DjCompHistConfig(AppConfig):
-    name = 'dj_comp_hist'
-
-    def ready(self):
-        import dj_comp_hist.signals
-
